refactor: Implement FSD structure with index files #78
