+++ apps/admin/src/app/index.ts
@@ -1,5 +1,3 @@
 // 애플리케이션 로직이 초기화되는 곳입니다. 프로바이더, 라우터, 전역 스타일, 전역 타입 선언 등이 여기에서 정의됩니다. 애플리케이션의 진입점 역할을 합니다.
 
-export { ReactQueryProvider } from './providers';
-// export { TrpcProvider } from './providers';
-export { createRouter } from './providers';
+export * from './providers';
