+++ apps/admin/src/entities/index.ts
@@ -1,2 +1,3 @@
 // 이 레이어는 비즈니스 엔티티를 나타냅니다. 이러한 엔티티에는 사용자, 리뷰, 댓글 등이 포함될 수 있습니다. 선택적 레이어입니다.
-export { type Post } from './post/model/post';
+
+export * from './post/model/post';
