+++ apps/admin/src/features/blog/model/index.ts
@@ -1,0 +1,2 @@
+export { usePostTable } from './usePostTable';
+export { usePostTableFilter } from './usePostTableFilter';
