--- apps/admin/src/features/blog/ui/PostEditor/ErrorMessage.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const errorMessage = style({
-	color: '#e53e3e',
-	fontSize: '0.875rem',
-	marginTop: '0.25rem',
-});
