+++ apps/admin/src/features/blog/ui/PostEditor/ErrorMessage.tsx
@@ -1,13 +1,12 @@
-import { Box } from '@jung/design-system/components';
+import { Typography } from '@jung/design-system/components';
 import type React from 'react';
-import * as styles from './ErrorMessage.css';
 
 interface ErrorMessageProps {
 	message: string;
 }
 
 const ErrorMessage: React.FC<ErrorMessageProps> = ({ message }) => (
-	<Box className={styles.errorMessage}>{message}</Box>
+	<Typography.SubText color='error'>{message}</Typography.SubText>
 );
 
 export default ErrorMessage;
