+++ apps/admin/src/features/blog/ui/PostTable/TableContent.tsx
@@ -1,5 +1,5 @@
+import { usePostTable } from '@/fsd/features/blog/model';
 import { Box } from '@jung/design-system/components';
-import { usePostTable } from '../../model/usePostTable';
 import ErrorFallback from '../ErrorFallback';
 import { TableBody } from './TableBody';
 import { TableHeader } from './TableHeader';
