+++ apps/admin/src/features/blog/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as PostEditor } from './PostEditor/PostEditor';
+export { default as PostTable } from './PostTable/PostTable';
