+++ apps/admin/src/features/index.ts
@@ -4,3 +4,5 @@ export { fetchPosts } from './blog/api/getPosts';
 
 export { useGetAllPosts as usePostsQuery } from './blog/api/useGetAllPosts';
 export type { PostFilters } from './blog/types/postFilters';
+
+export * from './blog';
