+++ apps/admin/src/shared/index.ts
@@ -1,8 +1,5 @@
-export { getPageTitle } from './lib';
-export { usePathname } from './lib/hooks';
+export * from './lib';
+
 export { Routes, RoutesArray } from './config/routes';
 export { supabase } from './api/supabase';
 export { postKeys } from '../features';
-export { useThrottle } from './lib/hooks/useThrottle';
-export { default as storage } from './lib/utils/localStorage';
-export { useErrorHandler } from './lib/hooks/useErrorHandler';
