--- apps/admin/src/shared/lib/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as usePathname } from './usePathname';
