+++ apps/admin/src/shared/lib/index.ts
@@ -1,1 +1,4 @@
 export { default as getPageTitle } from './utils/getPageTitle';
+export { default as usePathname } from './hooks/usePathname';
+export { useThrottle } from './hooks/useThrottle';
+export { default as storage } from './utils/localStorage';
