+++ apps/admin/src/views/blog/index.ts
@@ -1,0 +1,2 @@
+export { default as BlogPage } from './ui/BlogPage';
+export { default as PostFormPage } from './ui/PostFormPage';
