--- apps/admin/src/widgets/Header/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as Header } from './Header';
