+++ apps/admin/src/widgets/Layout/index.ts
@@ -1,0 +1,1 @@
+export { default as Layout } from './ui/Layout';
