--- apps/admin/src/widgets/Layout/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as Layout } from './Layout';
