+++ apps/admin/src/widgets/Sidebar/index.ts
@@ -1,0 +1,1 @@
+export { default as Sidebar } from './ui/Sidebar';
