--- apps/admin/src/widgets/Sidebar/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as Sidebar } from './Sidebar';
