+++ apps/admin/src/widgets/index.ts
@@ -1,3 +1,5 @@
 // 페이지에 사용되는 독립적인 UI 컴포넌트입니다.
 
-export { Layout } from './Layout/ui';
+export * from './Header';
+export * from './Sidebar';
+export * from './Layout';
